Add search, paginatiom, menu, login andtable for courses

+++ education-management-system/src/app/courses/courses.component.ts
@@ -1,10 +1,17 @@
 
-import { Component, OnInit } from '@angular/core';
+import { Component, OnInit, ViewChild } from '@angular/core';
 import { AppServiceService } from '../app-service.service';
 import { DialogInvComponent } from '../dialog-inv/dialog-inv.component';
 import { Cours } from './cours';
 import { MatDialog } from '@angular/material/dialog';
 import { FormControl, FormGroup, Validators } from '@angular/forms';
+import { MatPaginator, PageEvent } from '@angular/material/paginator';
+import { DialogDeleteComponent } from '../dialog-delete/dialog-delete.component';
+import { MatSort } from '@angular/material/sort';
+import { MatTableDataSource } from '@angular/material/table';
+import { SomeDataService } from '../some-data.service';
+
+
 
 @Component({
   selector: 'app-courses',
@@ -13,7 +20,10 @@ import { FormControl, FormGroup, Validators } from '@angular/forms';
 })
 export class CoursesComponent implements OnInit {
 
-  courses: Cours[] = [];
+
+    courses: Cours[] = [];
+  pageSlice: Cours[]=[];
+  len=0;
 
 
   role:string=""
@@ -21,6 +31,10 @@ export class CoursesComponent implements OnInit {
   roleForUsers:string=""
   authForUsers:string=""
   uCours:string="";
+
+  strsearch="";
+  
+
 
   dialog: any;
 
@@ -33,37 +47,48 @@ export class CoursesComponent implements OnInit {
   })
  
 
-  constructor(private service:AppServiceService,  public dialog1: MatDialog ) { }
+  constructor(private service:AppServiceService,  public dialog1: MatDialog , private someSrv: SomeDataService) { }
+
+ 
+
+  OnPageChange(event: PageEvent){
+    console.log(event);
+    const startIndex = event.pageIndex*event.pageSize;
+    let endIndex=startIndex+event.pageSize;
+    if(endIndex>this.courses.length){
+        endIndex=this.courses.length;
+    }
+    this.pageSlice=this.courses.slice(startIndex, endIndex);
+  }
 
   ngOnInit(): void {
-    this.role=String(localStorage.getItem('roleForE')).split('"').join('');
+
     this.auth=String(localStorage.getItem('logForE')).split('"').join('');
     this.uCours=String(localStorage.getItem('uCours')).split('"').join('');
 
-    let v11=(<HTMLInputElement>document.getElementById('test')).value;
+    this.role=this.someSrv.role;
     
 
+    this.service.getCourses()
+   .subscribe(courses => this.pageSlice=courses.slice(0,3) );
    
 
     
    
 
     if (this.role=="admin" || this.role=="rootadmin"){
-      (<HTMLInputElement>document.getElementById('vv33')).style.display="block";
-      (<HTMLInputElement>document.getElementById('v2')).style.display="block";
+
       this.uCours="";
       
     }
     if (this.role=="user"){
-      (<HTMLInputElement>document.getElementById('v2')).style.display="none";
-      (<HTMLInputElement>document.getElementById('vv33')).style.display="none";
+
       this.uCours=String(localStorage.getItem('uCours')).split('"').join('');
       
       
     }
     if (this.role=="manager"){
-      (<HTMLInputElement>document.getElementById('v2')).style.display="block";
-      (<HTMLInputElement>document.getElementById('vv33')).style.display="block";
+
       this.uCours="";
       
       this.roleForUsers=String(localStorage.getItem('roleForE')).split('"').join('');
@@ -101,10 +126,12 @@ this.getCourses();
     .subscribe(courses => {
 
       this.courses= courses;
+
+      this.len=this.courses.length;
     for( let i=0; i<this.courses.length; i++){
 
         if (Math.ceil( ((new Date(this.courses[i].test).getTime())-this.now.getTime())/ (1000 * 3600 * 24))<5 && Math.ceil( ((new Date(this.courses[i].test).getTime())-this.now.getTime())/ (1000 * 3600 * 24))>0){
-        alert("You have a test" ) ;
+      //  alert("You have a test" ) ;
         break;      
       }
       }
@@ -112,40 +139,59 @@ this.getCourses();
     });
 
 
+    this.service.getCourses()
+   .subscribe(courses => this.pageSlice=courses.slice(0,3) );
 
   }
 
   delete(cours: Cours): void {
+    let dr=this.dialog1.open(DialogDeleteComponent);
+    dr.afterClosed().subscribe((result: any) => {
+ if(result){
     this.courses = this.courses.filter(c => c !== cours);
     this.service.deleteCours(cours.id).subscribe();
+this.getCourses();
+
+   } })
+   
+   
+
   }
 
   
   add(): void {
-   let v1=(<HTMLInputElement>document.getElementById('v1')).style.display="block"
-   let v2=(<HTMLInputElement>document.getElementById('v2')).style.display="none"
-   let v4=(<HTMLInputElement>document.getElementById('v4')).style.display="block"
+
+
+   this.edited=1;
+   this.buttonAdd=1;
+   this.roundtn=0;
   
   }
 
 
   save(id1:string, name: string, plan: string, test:string): void {
     
-    let v1=(<HTMLInputElement>document.getElementById('v1')).style.display="none"
-    let v3=(<HTMLInputElement>document.getElementById('v3')).style.display="none"
-    let v5=(<HTMLInputElement>document.getElementById('v5')).style.display="block"
+ 
     const id=Number(id1);
     this.service.updateHero({ id, name, plan, test }  as Cours)
     .subscribe();
     this.getCourses();
+    this.edited=0;
+    this.buttonEdit=0;
+    this.roundtn=1;
+    
+this.getCourses();
+
+
   }
 
+  edited=0;
+  buttonEdit=0;
+  buttonAdd=0;
+  roundtn=1;
 
   savea(name: string, plan: string, test: string): void {
-    let v1=(<HTMLInputElement>document.getElementById('v1')).style.display="none"
-    let v2=(<HTMLInputElement>document.getElementById('v2')).style.display="block"
-    let v4=(<HTMLInputElement>document.getElementById('v4')).style.display="none"
-    let v5=(<HTMLInputElement>document.getElementById('v5')).style.display="block"
+
 
     this.auth=String(localStorage.getItem('logForE')).split('"').join('');
 
@@ -158,25 +204,31 @@ this.getCourses();
       .subscribe(cours => {
         this.courses.push(cours);
       });
+      this.edited=0;
+      this.buttonAdd=0;
+      this.roundtn=1;
+      
+this.getCourses();
   }
   
   
 
   edit(cours: Cours):Cours{
-    (<HTMLInputElement>document.getElementById('name')).value=cours.name;
+  /* (<HTMLInputElement>document.getElementById('name')).value=cours.name;
     (<HTMLInputElement>document.getElementById('plan')).value=cours.plan;
     (<HTMLInputElement>document.getElementById('test')).value=cours.test;
-    ((<HTMLInputElement>document.getElementById('id1')).value)=String(cours.id);
-    let v1=(<HTMLInputElement>document.getElementById('v1')).style.display="block"
-    let v4=(<HTMLInputElement>document.getElementById('v3')).style.display="block"
-    let v5=(<HTMLInputElement>document.getElementById('v5')).style.display="none"
-    
+    ((<HTMLInputElement>document.getElementById('id1')).value)=String(cours.id);*/
+  
+    this.edited=1;
+    this.buttonEdit=1;
+    this.roundtn=0;
     
     return cours;
   
 
   }
 
+    displayedColumns: string[] = [ 'name', 'plan', 'test', 'menu'];
 
 
 }
